feat: using ForceDeleteNodes to remove unregistered nodes and removing unregistered nodes deletion blockers

+++ cluster-autoscaler/core/static_autoscaler.go
@@ -751,22 +751,7 @@ func (a *StaticAutoscaler) removeOldUnregisteredNodes(allUnregisteredNodes []clu
 	removedAny := false
 	for nodeGroupId, unregisteredNodesToDelete := range nodesToBeDeletedByNodeGroupId {
 		nodeGroup := nodeGroups[nodeGroupId]
-
-		klog.V(0).Infof("Removing %v unregistered nodes for node group %v", len(unregisteredNodesToDelete), nodeGroupId)
-		size, err := nodeGroup.TargetSize()
-		if err != nil {
-			klog.Warningf("Failed to get node group size; nodeGroup=%v; err=%v", nodeGroup.Id(), err)
-			continue
-		}
-		possibleToDelete := size - nodeGroup.MinSize()
-		if possibleToDelete <= 0 {
-			klog.Warningf("Node group %s min size reached, skipping removal of %v unregistered nodes", nodeGroupId, len(unregisteredNodesToDelete))
-			continue
-		}
-		if len(unregisteredNodesToDelete) > possibleToDelete {
-			klog.Warningf("Capping node group %s unregistered node removal to %d nodes, removing all %d would exceed min size constaint", nodeGroupId, possibleToDelete, len(unregisteredNodesToDelete))
-			unregisteredNodesToDelete = unregisteredNodesToDelete[:possibleToDelete]
-		}
+		
 		nodesToDelete := toNodes(unregisteredNodesToDelete)
 
 		opts, err := nodeGroup.GetOptions(a.NodeGroupDefaults)
@@ -784,8 +769,15 @@ func (a *StaticAutoscaler) removeOldUnregisteredNodes(allUnregisteredNodes []clu
 			}
 			nodesToDelete = instancesToFakeNodes(instances)
 		}
+		
 
-		err = nodeGroup.DeleteNodes(nodesToDelete)
+		// CloudProviders that have implemented ForceDeleteNodes will force remove unregistered nodes
+		err = nodeGroup.ForceDeleteNodes(nodesToDelete) 
+		if err == cloudprovider.ErrNotImplemented {
+			// CloudProviders that have checks for min size and don't have ForceDeleteNodes implemented will 
+			// still contain unregistered nodes if that deletion violates min count.
+			err = nodeGroup.DeleteNodes(nodesToDelete)
+		}
 		csr.InvalidateNodeInstancesCacheEntry(nodeGroup)
 		if err != nil {
 			klog.Warningf("Failed to remove %v unregistered nodes from node group %s: %v", len(nodesToDelete), nodeGroupId, err)
